requirements.txt
psycopg2-binary
fastapi[standard]
uvicorn
python-dotenv
pydantic 